% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize_authors.R
\name{capitalize_authors}
\alias{capitalize_authors}
\title{Capitalize author names}
\usage{
capitalize_authors(x, make.lower = c("De", "Da", "Do", "E"))
}
\arguments{
\item{x}{data frame containing a column "Authors", author names (separated by commas, if multiple)}

\item{make.lower}{character vector containing name conectors to keep in lowercase}
}
\value{
data frame with (hopefully) correctly capitalized author names in column "Authors".
}
\description{
Performs capitalization of author names
}
\details{
This function attempts to standardize capitalization of author names for the production listings
}
