% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize_titles.R
\name{capitalize_titles}
\alias{capitalize_titles}
\title{Capitalize titles}
\usage{
capitalize_titles(x, which.fields = c("Title", "Bookname", "Journal",
  "Conference"))
}
\arguments{
\item{x}{dataframe containing fields to be capitalized (using sentence case)}

\item{which.fields}{character vector with names of dataframe columns to be
capitalized}
}
\value{
capitalized string
}
\description{
Performs capitalization of titles
}
\details{
This function attempts to standardize capitalization of titles for the
productions listing, using sentence case.
}
