% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_accepted_papers.R
\name{get_accepted_papers}
\alias{get_accepted_papers}
\title{Extract accepted journal papers}
\usage{
get_accepted_papers(x, ID = stats::runif(1))
}
\arguments{
\item{x}{Lattes list (generated internally in \code{\link[=lattes_to_list]{lattes_to_list()}})}

\item{ID}{a unique identifier for each CV being processed.}
}
\value{
data frame containing parsed information on accepted journal papers
}
\description{
Extracts accepted journal papers from Lattes list.
}
\details{
This function extracts relevant information on accepted journal papers from a Lattes list.
}
