% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_authors.R
\name{get_authors}
\alias{get_authors}
\title{Extract author names}
\usage{
get_authors(item)
}
\arguments{
\item{x}{list generated by \code{\link[=Lattes_to_list]{Lattes_to_list()}}}

\item{ID}{a unique identifier for each CV being processed.}
}
\value{
character string containing author names in "name surname" format,
separated by commas.
}
\description{
Extracts author names from items
}
\details{
This function extracts relevant information on accepted journal papers from a list generated by \code{\link[=Lattes_to_list]{Lattes_to_list()}}.
}
