% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_books.R
\name{get_books}
\alias{get_books}
\title{Extract published books}
\usage{
get_books(x, ID = stats::runif(1))
}
\arguments{
\item{x}{Lattes list (generated internally in \code{\link[=lattes_to_list]{lattes_to_list()}})}

\item{ID}{a unique identifier for each CV being processed.}
}
\value{
data frame containing parsed information on published books
}
\description{
Extracts published books from Lattes list
}
\details{
This function extracts relevant information on published books from a Lattes list
}
