% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_journal_papers.R
\name{get_journal_papers}
\alias{get_journal_papers}
\title{Extract published journal papers}
\usage{
get_journal_papers(x, ID = stats::runif(1))
}
\arguments{
\item{x}{Lattes list (generated internally in \code{\link[=lattes_to_list]{lattes_to_list()}})}

\item{ID}{a unique identifier for each CV being processed.}
}
\value{
data frame containing parsed information on journal papers
}
\description{
Extracts published journal papers from Lattes list.
}
\details{
This function extracts relevant information on published journal papers from a Lattes list.
}
