% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattes_to_list.R
\name{lattes_to_list}
\alias{lattes_to_list}
\title{Convert a set of Lattes CV XML files to a list object}
\usage{
lattes_to_list(filenames = NULL, CV.dir = NULL)
}
\arguments{
\item{filenames}{character vector containing name of the file(s) to be
processed. If NULL then use all XML files in the folder given by \code{CV.dir}}

\item{CV.dir}{folder where CVs are contained, relative to current working
directory (\strong{do not} use "./" to refer the the current folder). If NULL
then the current working directory is used.}
}
\value{
list vector where each element is a dataframe with information on a
specific aspect of the academic production
}
\description{
Extract information from a set of Lattes XML files and convert it to a list
vector
}
\details{
This function extracts relevant information from a set of Lattes CV XML files
and outputs a list object containing specific information on the following
aspects of a group's production:
\itemize{
\item Accepted journal papers
\item Published journal papers
\item Published conference papers
\item Published book chapters
\item Published books
\item Ph.D. student defenses
\item M.Sc. student defenses
}

Journal and conference papers are checked for duplication using DOI and Title
information. Duplicated entries are registered only once.
}
