% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_name_surname.R
\name{make_name_surname}
\alias{make_name_surname}
\title{Put author names in "Name Surname" format}
\usage{
make_name_surname(x)
}
\arguments{
\item{x}{string containing author name, either in "Surname, Name" or "Name Surname" format}
}
\value{
string containing author name in "Name Surname" format
}
\description{
Put author names in "Name Surname" format
}
