% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_accepted.R
\name{print_accepted}
\alias{print_accepted}
\title{Print accepted papers}
\usage{
print_accepted(x, language = c("EN", "PT"))
}
\arguments{
\item{x}{data frame containing information on accepted papers}

\item{language}{Language to use in section headers}
}
\description{
Prints accepted papers
}
