% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_book_chapters.R
\name{print_book_chapters}
\alias{print_book_chapters}
\title{Print book chapters}
\usage{
print_book_chapters(x)
}
\arguments{
\item{x}{data frame containing information on published book chapters}
}
\description{
Prints published book chapters
}
