% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_books.R
\name{print_books}
\alias{print_books}
\title{Print books}
\usage{
print_books(x, language = c("EN", "PT"))
}
\arguments{
\item{x}{data frame containing information on published books}

\item{language}{Language to use in section headers}
}
\description{
Prints published books
}
