% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_journal_papers.R
\name{print_journal_papers}
\alias{print_journal_papers}
\title{Print journal papers}
\usage{
print_journal_papers(x)
}
\arguments{
\item{x}{data frame containing information on published papers}
}
\description{
Prints published journal papers
}
