% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_msc_theses.R
\name{print_msc_theses}
\alias{print_msc_theses}
\title{Print MSc theses}
\usage{
print_msc_theses(x, language = c("EN", "PT"))
}
\arguments{
\item{x}{data frame containing information on published MSc theses}

\item{language}{Language to use in section headers}
}
\description{
Prints MSc theses defended
}
