% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_phd_theses.R
\name{print_phd_theses}
\alias{print_phd_theses}
\title{Print PhD theses}
\usage{
print_phd_theses(x, language = c("EN", "PT"))
}
\arguments{
\item{x}{data frame containing information on published phd theses}

\item{language}{Language to use in section headers}
}
\description{
Prints PhD theses defended
}
