% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Function to remove duplicated entries}
\usage{
remove_duplicates(x)
}
\arguments{
\item{x}{data frame containing information on a certain type of production (journal papers, conference papers, etc.).
Must contain columns "Title" and "Year". Optional (very useful) columns include "DOI" and "ISBN".}
}
\value{
data frame with duplicates removed
}
\description{
Cleans up duplicated items on a production data frame based on DOI, ISBN, and Title.
This function also removes any rows where the value of "year" is greater than the current year (since these are considered errors)
}
