% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_papers.R
\name{sort_papers}
\alias{sort_papers}
\title{Sort publications by year}
\usage{
sort_papers(x, decreasing = TRUE)
}
\arguments{
\item{x}{data frame containing information on a certain type of production (journal papers, conference papers, etc.).
Must contain column "Year". Optional (useful) columns include "DOI" and "ISBN".}

\item{decreasing}{flag, sort in decreasing or increasing order}
}
\value{
ordered data frame
}
\description{
Sorts publications by year
}
