% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize_authors.R
\name{capitalize_authors}
\alias{capitalize_authors}
\title{Capitalize author names}
\usage{
capitalize_authors(x, make.lower = c("De", "Di", "Do", "Da", "Dos", "Das",
  "Dello", "Della", "Dalla", "Dal", "Del", "E", "Em", "Na", "No", "Nas", "Nos",
  "Van", "Von", "Y"), author.aliases = NULL)
}
\arguments{
\item{x}{data frame containing a column "Authors" (or "Author"),
which should describe author names (separated by commas, if multiple) for a given
production type.}

\item{make.lower}{character vector containing name conectors to keep in lowercase}

\item{author.aliases}{list vector with author aliases.
See \code{\link[=lattes_to_list]{lattes_to_list()}} for details.}
}
\value{
data frame with (hopefully) correctly capitalized author names in column "Authors".
}
\description{
Performs capitalization of author names
}
\details{
This function attempts to standardize capitalization of author names for the production listings
}
