% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_authors.R
\name{get_authors}
\alias{get_authors}
\title{Extract author names}
\usage{
get_authors(item)
}
\arguments{
\item{item}{a single bibliographic item, selected from a lattes list object.}
}
\value{
character string containing author names in "name surname" format,
separated by commas.
}
\description{
Extracts author names from items
}
\details{
This function extracts author names from bibliographic items.
}
