% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_conferences.R
\name{print_conferences}
\alias{print_conferences}
\title{Print conference papers}
\usage{
print_conferences(x, isIntl = TRUE, language = c("EN", "PT"))
}
\arguments{
\item{x}{data frame containing information on published conference papers}

\item{isIntl}{logical flag, TRUE for international conferences, FALSE for
national/regional}

\item{language}{Language to use in section headers}
}
\description{
Prints published conference papers
}
