% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_journal_papers.R
\name{print_journal_papers}
\alias{print_journal_papers}
\title{Print journal papers}
\usage{
print_journal_papers(x, language = c("EN", "PT"))
}
\arguments{
\item{x}{data frame containing information on published papers}

\item{language}{Language to use in section headers}
}
\description{
Prints published journal papers
}
